Move implementation of get_window_around_maximum to NuRadioReco.utilities.diodeSimulator.get_window_around_maximum
