+++ NuRadioReco/examples/AliasPhasedArray/SNR_study/T02SNRNyquist.py
@@ -52,7 +52,7 @@ import NuRadioReco.modules.channelGenericNoiseAdder
 import NuRadioReco.utilities.diodeSimulator
 
 
-from NuRadioReco.utilities.traceWindows import get_window_around_maximum
+from NuRadioReco.utilities.diodeSimulator import get_window_around_maximum
 from NuRadioReco.utilities import units
 from NuRadioMC.simulation import simulation
 import NuRadioReco.modules.analogToDigitalConverter
