+++ NuRadioReco/examples/EnvelopePhasedArray/Effective_volume/T02RunPhasedRNO_shortband.py
@@ -30,7 +30,7 @@ import NuRadioReco.modules.channelGenericNoiseAdder
 from NuRadioReco.utilities import units
 from NuRadioMC.simulation import simulation
 import NuRadioReco.utilities.diodeSimulator
-from NuRadioReco.utilities.traceWindows import get_window_around_maximum
+from NuRadioReco.utilities.diodeSimulator import get_window_around_maximum
 import numpy as np
