+++ NuRadioReco/utilities/_deprecated/__init__.py
@@ -1,4 +1,5 @@
 __all__ = [
     'bandpass_filter',
-    'variableWindowSizeCorrelation'
+    'variableWindowSizeCorrelation',
+    'traceWindows',
 ]
